Updated codebase to be inline with the scaffold repo and fixed coding standards.
