+++ .editorconfig
@@ -1,8 +1,18 @@
+# This is the top-most .editorconfig file; do not search in parent directories.
+root = true
+
+# All files.
+[*]
+end_of_line = LF
+indent_style = space
+indent_size = 2
+charset = utf-8
+trim_trailing_whitespace = true
+insert_final_newline = true
+
 [*.php]
 indent_size = 2
 
-[*.js]
-indent_size = 2
 
 [*.{json,lock}]
 indent_size = 4
