+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,16 @@
+## Checklist before requesting a review
+
+- [ ] I have provided information in `Changed` section about WHY something was
+  done if this was not a normal implementation
+- [ ] I have performed a self-review of my code
+- [ ] I have commented my code, particularly in hard-to-understand areas
+- [ ] I have added tests that prove my fix is effective or that my feature works
+- [ ] I have run new and existing relevant tests locally with my changes, and
+  they passed
+- [ ] I have provided screenshots, where applicable
+
+## Changed
+
+1.
+
+## Screenshots
