+++ .github/workflows/auto-assign-pr-author.yml
@@ -2,7 +2,9 @@ name: 'Auto Author Assign'
 
 on:
   pull_request_target:
-    types: [opened, reopened]
+    types:
+      - opened
+      - reopened
 
 permissions:
   pull-requests: write
@@ -11,4 +13,4 @@ jobs:
   assign-author:
     runs-on: ubuntu-latest
     steps:
-      - uses: toshimaru/auto-author-assign@v1.4.0
+      - uses: toshimaru/auto-author-assign@v1.6.2
