+++ .github/workflows/test.yml
@@ -1,31 +1,32 @@
 name: Test
 
 on:
-  push:
+  pull_request:
     branches:
       - main
       - 'feature/**'
+  push:
+    branches:
+      - main
 
 jobs:
   test-php:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@v4
     - uses: actions/cache@v3
       with:
         path: /tmp/composer-cache
         key: ${{ runner.os }}-${{ hashFiles('**/composer.lock') }}
     - uses: php-actions/composer@v6
-    - run: composer lint || true
-    - run: composer test
-    - uses: paambaati/codeclimate-action@v5.0.0
-      env:
-        CC_TEST_REPORTER_ID: ${{secrets.CC_TEST_REPORTER_ID}}
+    - run: composer lint
+    - run: composer test:coverage
+    - uses: insightsengineering/coverage-action@v2
       with:
-        coverageCommand: composer test:coverage
-        coverageLocations: |
-          ${{github.workspace}}/.coverage-cobertura.xml:cobertura
-    - uses: actions/upload-artifact@v3
-      with:
-        name: code-coverage-report
-        path: .coverage-html
+        path: cobertura.xml
+        publish: true
+        diff: true
+        threshold: 90
+        fail: true
+        coverage-reduction-failure: true
+        new-uncovered-statements-failure: true
