+++ README.md
@@ -1,18 +1,31 @@
-# shell-var-lint
+<p align="center">
+  <a href="" rel="noopener">
+  <img width=200px height=200px src="https://placehold.jp/000000/ffffff/200x200.png?text=Shell%20var%20lint&css=%7B%22border-radius%22%3A%22%20100px%22%7D" alt="Project logo"></a>
+</p>
 
+<h1 align="center">shell-var-lint</h1>
 
-[![Test](https://github.com/alexskrypnyk/shell-var-lint/actions/workflows/test.yml/badge.svg)](https://github.com/alexskrypnyk/shell-var-lint/actions/workflows/test.yml)
-![GitHub release (latest by date)](https://img.shields.io/github/v/release/alexskrypnyk/shell-var-lint)
-![LICENSE](https://img.shields.io/github/license/alexskrypnyk/shell-var-lint)
-[![Test Coverage](https://api.codeclimate.com/v1/badges/ab788250b78c9c84600c/test_coverage)](https://codeclimate.com/github/AlexSkrypnyk/shell-var-lint/test_coverage)
-[![Maintainability](https://api.codeclimate.com/v1/badges/ab788250b78c9c84600c/maintainability)](https://codeclimate.com/github/AlexSkrypnyk/shell-var-lint/maintainability)
+<div align="center">
 
+  [![GitHub Issues](https://img.shields.io/github/issues/AlexSkrypnyk/shell-var-lint.svg)](https://github.com/AlexSkrypnyk/shell-var-lint/issues)
+  [![GitHub Pull Requests](https://img.shields.io/github/issues-pr/AlexSkrypnyk/shell-var-lint.svg)](https://github.com/AlexSkrypnyk/shell-var-lint/pulls)
+  [![Test](https://github.com/AlexSkrypnyk/shell-var-lint/actions/workflows/test.yml/badge.svg)](https://github.com/AlexSkrypnyk/shell-var-lint/actions/workflows/test.yml)
+  ![Coverage](https://github.com/AlexSkrypnyk/shell-var-lint/blob/_xml_coverage_reports/data/main/badge.svg)
+  ![GitHub release (latest by date)](https://img.shields.io/github/v/release/AlexSkrypnyk/shell-var-lint)
+  ![LICENSE](https://img.shields.io/github/license/AlexSkrypnyk/shell-var-lint)
+
+</div>
+
+---
+
+<p align="center"> Lint and fix shell vars to ${VAR} format.
+    <br>
+</p>
 
 ## Features
 
-- Your first feature as a list item
-- Your second feature as a list item
-- Your third feature as a list item
+- Report on shell variables that are not in `${VAR}` format.
+- Fix shell variables that are not in `${VAR}` format.
 
 ## Installation
