+++ phpcs.xml
@@ -11,11 +11,24 @@
     <arg value="p"/>
     <arg name="colors"/>
 
+    <!--
+    PHP_CodeSniffer does not support scanning of files without extensions, so
+    we have to use the ugly workaround to copy the file into a file with a 'php'
+    extension, scan that file, and then delete it.
+    @see https://github.com/squizlabs/PHP_CodeSniffer/issues/2916
+    -->
     <file>shell-var-lint</file>
     <file>tests/phpunit</file>
 
     <!-- Allow long array lines in tests. -->
     <rule ref="Drupal.Arrays.Array.LongLineDeclaration">
+        <exclude-pattern>*.Test\.php</exclude-pattern>
+        <exclude-pattern>*.TestCase\.php</exclude-pattern>
+        <exclude-pattern>*.test</exclude-pattern>
+    </rule>
+
+    <!-- Allow missing function names in tests. -->
+    <rule ref="Drupal.Commenting.FunctionComment.Missing">
         <exclude-pattern>*.Test\.php</exclude-pattern>
         <exclude-pattern>*.TestCase\.php</exclude-pattern>
         <exclude-pattern>*.test</exclude-pattern>
