+++ phpstan.neon
@@ -4,7 +4,7 @@
 
 parameters:
 
-  level: 1
+  level: 9
 
   paths:
     - shell-var-lint
@@ -14,5 +14,10 @@ parameters:
     - vendor/*
     - node_modules/*
 
-  # ignoreErrors:
-  #  - "#^Error message#"
+  ignoreErrors:
+    -
+      # Since tests and data providers do not have to have parameter docblocks,
+      # it is not possible to specify the type of the parameter, so we ignore
+      # this error.
+      message: '#.*no value type specified in iterable type array.#'
+      path: tests/phpunit/*
